+++ src/main/java/com/springcooler/sgma/studygroup/command/domain/service/DomainStudyGroupService.java
@@ -1,13 +1,7 @@
 package com.springcooler.sgma.studygroup.command.domain.service;
 
-import com.springcooler.sgma.studygroup.command.domain.aggregate.StudyGroupStatus;
-import org.springframework.stereotype.Service;
+public interface DomainStudyGroupService {
 
-@Service
-public class DomainStudyGroupService {
-
-    public boolean isActive(String groupStatus) {
-        return groupStatus.equals(StudyGroupStatus.ACTIVE.name());
-    }
+    boolean isActive(String groupStatus);
 
 }
