+++ src/main/java/com/springcooler/sgma/studygroup/command/infrastructure/service/InfraStudyGroupService.java
@@ -1,4 +1,4 @@
 package com.springcooler.sgma.studygroup.command.infrastructure.service;
 
-public class InfraStudyGroupService {
+public interface InfraStudyGroupService {
 }
